fix process close
