+++ src/api/cancel.go
@@ -11,6 +11,16 @@ import (
 // Method: PUT
 func cancel(gateway Gatewayer) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
+		// allow only one request at a time
+		closeFunc, err := serialize(gateway)
+		if err != nil {
+			logger.Error("serialize failed: %s", err.Error())
+			resp := NewHTTPErrorResponse(http.StatusInternalServerError, err.Error())
+			writeHTTPResponse(w, resp)
+			return
+		}
+		defer closeFunc()
+
 		if r.Method != http.MethodPut {
 			resp := NewHTTPErrorResponse(http.StatusMethodNotAllowed, "")
 			writeHTTPResponse(w, resp)
@@ -49,7 +59,7 @@ func cancel(gateway Gatewayer) http.HandlerFunc {
 				Data: failureMsg,
 			})
 		} else {
-			HandleFirmwareResponseMessages(w, r, gateway, msg)
+			HandleFirmwareResponseMessages(w, gateway, msg)
 		}
 	}
 }
