+++ src/api/http.go
@@ -1,7 +1,6 @@
 package api
 
 import (
-	"encoding/json"
 	"fmt"
 	"net"
 	"net/http"
@@ -10,11 +9,8 @@ import (
 	"time"
 
 	"github.com/NYTimes/gziphandler"
-	"github.com/gogo/protobuf/proto"
 	"github.com/rs/cors"
 	deviceWallet "github.com/skycoin/hardware-wallet-go/src/device-wallet"
-	"github.com/skycoin/hardware-wallet-go/src/device-wallet/wire"
-	messages "github.com/skycoin/hardware-wallet-protob/go"
 	wh "github.com/skycoin/skycoin/src/util/http"
 	"github.com/skycoin/skycoin/src/util/logging"
 )
@@ -34,6 +30,9 @@ const (
 
 var (
 	logger = logging.MustGetLogger("daemon-api")
+
+	// custom lock to help with serializing requests
+	ongoingOperation chan struct{}
 )
 
 // corsRegex matches all localhost origin headers
@@ -50,11 +49,13 @@ func init() {
 		logger.Panic(err)
 	}
 
+	// set size to 1 to allow only 1 request at a time
+	ongoingOperation = make(chan struct{}, 1)
+
 	apb := os.Getenv("AUTO_PRESS_BUTTONS")
 	if apb == "1" {
 		autoPressEmulatorButtons = true
 	}
-
 }
 
 type muxConfig struct {
@@ -81,63 +82,6 @@ type Config struct {
 	WriteTimeout       time.Duration
 	IdleTimeout        time.Duration
 	Mode               deviceWallet.DeviceType
-}
-
-// HTTPResponse represents the http response struct
-type HTTPResponse struct {
-	Error *HTTPError  `json:"error,omitempty"`
-	Data  interface{} `json:"data,omitempty"`
-}
-
-// ReceivedHTTPResponse parsed is a Parsed HTTPResponse
-type ReceivedHTTPResponse struct {
-	Error *HTTPError      `json:"error,omitempty"`
-	Data  json.RawMessage `json:"data"`
-}
-
-// HTTPError is included in an HTTPResponse
-type HTTPError struct {
-	Message string `json:"message"`
-	Code    int    `json:"code"`
-}
-
-// NewHTTPErrorResponse returns an HTTPResponse with the Error field populated
-func NewHTTPErrorResponse(code int, msg string) HTTPResponse {
-	if msg == "" {
-		msg = http.StatusText(code)
-	}
-
-	return HTTPResponse{
-		Error: &HTTPError{
-			Code:    code,
-			Message: msg,
-		},
-	}
-}
-
-func writeHTTPResponse(w http.ResponseWriter, resp HTTPResponse) {
-	out, err := json.MarshalIndent(resp, "", "    ")
-	if err != nil {
-		wh.Error500(w, "json.MarshalIndent failed")
-		return
-	}
-
-	w.Header().Add("Content-Type", ContentTypeJSON)
-
-	if resp.Error == nil {
-		w.WriteHeader(http.StatusOK)
-	} else {
-		if resp.Error.Code < 400 || resp.Error.Code >= 600 {
-			logger.Critical().Errorf("writeHTTPResponse invalid error status code: %d", resp.Error.Code)
-			w.WriteHeader(http.StatusInternalServerError)
-		} else {
-			w.WriteHeader(resp.Error.Code)
-		}
-	}
-
-	if _, err := w.Write(out); err != nil {
-		logger.WithError(err).Error("http Write failed")
-	}
 }
 
 // Serve serves the web interface on the configured host
@@ -274,11 +218,11 @@ func newServerMux(c muxConfig, gateway Gatewayer) *http.ServeMux {
 		}
 
 		handler = gziphandler.GzipHandler(handler)
+
 		mux.Handle(endpoint, handler)
 	}
 
 	webHandler := func(endpoint string, handler http.Handler) {
-		handler = wh.ElapsedHandler(logger, handler)
 		webHandlerWithOptionals(endpoint, handler, c.enableCSRF, !c.disableHeaderCheck)
 	}
 
@@ -321,220 +265,4 @@ func newServerMux(c muxConfig, gateway Gatewayer) *http.ServeMux {
 	webHandlerV1("/intermediate/word", wordRequestHandler(gateway))
 
 	return mux
-}
-
-// HandleFirmwareResponseMessages handles response messages from the firmware
-func HandleFirmwareResponseMessages(w http.ResponseWriter, r *http.Request, gateway Gatewayer, msg wire.Message) {
-	for {
-		switch msg.Kind {
-		case uint16(messages.MessageType_MessageType_PinMatrixRequest):
-			writeHTTPResponse(w, HTTPResponse{
-				Data: "PinMatrixRequest",
-			})
-			return
-		case uint16(messages.MessageType_MessageType_PassphraseRequest):
-			writeHTTPResponse(w, HTTPResponse{
-				Data: "PassPhraseRequest",
-			})
-			return
-		case uint16(messages.MessageType_MessageType_WordRequest):
-			writeHTTPResponse(w, HTTPResponse{
-				Data: "WordRequest",
-			})
-			return
-		case uint16(messages.MessageType_MessageType_ButtonRequest):
-			var err error
-			msg, err = gateway.ButtonAck()
-			if err != nil {
-				logger.Error(err.Error())
-				resp := NewHTTPErrorResponse(http.StatusUnauthorized, err.Error())
-				writeHTTPResponse(w, resp)
-			}
-		case uint16(messages.MessageType_MessageType_Failure):
-			failureMsg, err := deviceWallet.DecodeFailMsg(msg)
-			if err != nil {
-				resp := NewHTTPErrorResponse(http.StatusInternalServerError, err.Error())
-				writeHTTPResponse(w, resp)
-				return
-			}
-			resp := NewHTTPErrorResponse(http.StatusConflict, failureMsg)
-			writeHTTPResponse(w, resp)
-			return
-		case uint16(messages.MessageType_MessageType_Success):
-			successMsg, err := deviceWallet.DecodeSuccessMsg(msg)
-			if err != nil {
-				resp := NewHTTPErrorResponse(http.StatusUnauthorized, err.Error())
-				writeHTTPResponse(w, resp)
-				return
-			}
-
-			writeHTTPResponse(w, HTTPResponse{
-				Data: successMsg,
-			})
-			return
-
-		// AddressGen Response
-		case uint16(messages.MessageType_MessageType_ResponseSkycoinAddress):
-			addresses, err := deviceWallet.DecodeResponseSkycoinAddress(msg)
-			if err != nil {
-				resp := NewHTTPErrorResponse(http.StatusInternalServerError, err.Error())
-				writeHTTPResponse(w, resp)
-				return
-			}
-
-			writeHTTPResponse(w, HTTPResponse{
-				Data: addresses,
-			})
-			return
-
-		// Features Response
-		case uint16(messages.MessageType_MessageType_Features):
-			features := &messages.Features{}
-			err := proto.Unmarshal(msg.Data, features)
-			if err != nil {
-				resp := NewHTTPErrorResponse(http.StatusInternalServerError, err.Error())
-				writeHTTPResponse(w, resp)
-				return
-			}
-
-			writeHTTPResponse(w, HTTPResponse{
-				Data: features,
-			})
-			return
-
-		// SignMessage Response
-		case uint16(messages.MessageType_MessageType_ResponseSkycoinSignMessage):
-			signature, err := deviceWallet.DecodeResponseSkycoinSignMessage(msg)
-			if err != nil {
-				resp := NewHTTPErrorResponse(http.StatusInternalServerError, err.Error())
-				writeHTTPResponse(w, resp)
-				return
-			}
-
-			writeHTTPResponse(w, HTTPResponse{
-				Data: signature,
-			})
-			return
-
-		// TransactionSign Response
-		case uint16(messages.MessageType_MessageType_ResponseTransactionSign):
-			signatures, err := deviceWallet.DecodeResponseTransactionSign(msg)
-			if err != nil {
-				resp := NewHTTPErrorResponse(http.StatusInternalServerError, err.Error())
-				writeHTTPResponse(w, resp)
-				return
-			}
-
-			writeHTTPResponse(w, HTTPResponse{
-				Data: &signatures,
-			})
-			return
-		default:
-			resp := NewHTTPErrorResponse(http.StatusInternalServerError, fmt.Sprintf("recevied unexpected response message type: %s", messages.MessageType(msg.Kind)))
-			writeHTTPResponse(w, resp)
-			return
-		}
-	}
-}
-
-// PinMatrixRequest request data from /api/v1/intermediate/pin_matrix
-type PinMatrixRequest struct {
-	Pin string `json:"pin"`
-}
-
-func pinMatrixRequestHandler(gateway Gatewayer) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		if r.Method != http.MethodPost {
-			resp := NewHTTPErrorResponse(http.StatusMethodNotAllowed, "")
-			writeHTTPResponse(w, resp)
-			return
-		}
-
-		var req PinMatrixRequest
-		if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-			resp := NewHTTPErrorResponse(http.StatusBadRequest, err.Error())
-			writeHTTPResponse(w, resp)
-			return
-		}
-		defer r.Body.Close()
-
-		msg, err := gateway.PinMatrixAck(req.Pin)
-		if err != nil {
-			resp := NewHTTPErrorResponse(http.StatusInternalServerError, err.Error())
-			writeHTTPResponse(w, resp)
-			return
-		}
-
-		HandleFirmwareResponseMessages(w, r, gateway, msg)
-	}
-}
-
-// PassPhraseRequest request data from /api/v1/intermediate/passphrases
-type PassPhraseRequest struct {
-	Passphrase string `json:"passphrase"`
-}
-
-func passphraseRequestHandler(gateway Gatewayer) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		if r.Method != http.MethodPost {
-			resp := NewHTTPErrorResponse(http.StatusMethodNotAllowed, "")
-			writeHTTPResponse(w, resp)
-			return
-		}
-
-		var req PassPhraseRequest
-		if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-			resp := NewHTTPErrorResponse(http.StatusBadRequest, err.Error())
-			writeHTTPResponse(w, resp)
-			return
-		}
-		defer r.Body.Close()
-
-		msg, err := gateway.PassphraseAck(req.Passphrase)
-		if err != nil {
-			resp := NewHTTPErrorResponse(http.StatusInternalServerError, err.Error())
-			writeHTTPResponse(w, resp)
-		}
-
-		HandleFirmwareResponseMessages(w, r, gateway, msg)
-	}
-}
-
-// WordRequest request data from /api/v1/intermediate/word
-type WordRequest struct {
-	Word string `json:"word"`
-}
-
-func wordRequestHandler(gateway Gatewayer) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		if r.Method != http.MethodPost {
-			resp := NewHTTPErrorResponse(http.StatusMethodNotAllowed, "")
-			writeHTTPResponse(w, resp)
-			return
-		}
-
-		var req WordRequest
-		if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-			resp := NewHTTPErrorResponse(http.StatusBadRequest, err.Error())
-			writeHTTPResponse(w, resp)
-			return
-		}
-		defer r.Body.Close()
-
-		msg, err := gateway.WordAck(req.Word)
-		if err != nil {
-			resp := NewHTTPErrorResponse(http.StatusInternalServerError, err.Error())
-			writeHTTPResponse(w, resp)
-		}
-
-		HandleFirmwareResponseMessages(w, r, gateway, msg)
-	}
-}
-
-func newStrPtr(s string) *string {
-	return &s
-}
-
-func newUint32Ptr(n uint32) *uint32 {
-	return &n
 }
